pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "celadon"
description = 'A state-driven UI library for the terminal.'
readme = "README.md"
requires-python = ">=3.8"
keywords = []
authors = [
  { name = "bczsalba", email = "bczsalba@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "sh40-slate",
    "sh40-zenith",
    "sh40-carmine",
    "PyYAML",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/shade40/celadon#readme"
Issues = "https://github.com/shade40/celadon/issues"
Source = "https://github.com/shade40/celadon"

[project.scripts]
celadon = "celadon.__main__:main"

[tool.hatch.version]
path = "celadon/__about__.py"

[tool.hatch.build]
include = [
  "celadon/*.py",
  "celadon/py.typed",
  "/tests",
]

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
  "sh40-slate",
  "sh40-zenith",
  "sh40-carmine",
  "PyYAML",
]

[tool.hatch.envs.test]
dependencies = [
  "pytest",
  "pytest-cov",
  "sh40-slate",
  "sh40-zenith",
  "sh40-carmine",
  "PyYAML",
]

[tool.hatch.envs.default.scripts]
test = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=celadon --cov=tests && coverage html"
type = "mypy celadon"
upload = "hatch build && twine upload dist/* && hatch clean"

[[tool.hatch.envs.test.matrix]]
python = ["38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "celadon/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no-cov",
  "def __repr__",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
